clean unwanted comments and code content

+++ app/app.py
@@ -4,6 +4,7 @@ from core import Core
 from docopt import docopt, DocoptExit
 import templating
 import cmd
+import click
 
 
 templating.app_intro()
@@ -11,11 +12,17 @@ templating.intro_header()
 core = Core()
 
 class App(cmd.Cmd):
-    prompt = "-> "
+    prompt = "Amity->>> "
     # commads
 
     def do_help(self,  args):
       core.help()
+
+    def do_load_state(self,  args):
+      core.load_state()
+
+    def do_save_state(self,  args):
+      core.save_state()
 
     def do_add_room(self,  args):
         """
@@ -59,9 +66,17 @@ class App(cmd.Cmd):
 
         room_name = parsed_input["<name>"]
         if room_name:
-          core.delete_room(room_name)
+          response = click.secho("Are you sure you want to delete this room [y/n] n ?", fg='yellow', bold=True)
+          if response == "n" or response == "N":
+            click.secho("Delete Cancelled", fg='cyan', bold=True)
+          elif response == "y" or response == "Y": 
+            core.delete_room(room_name)
+          else:
+            click.secho("Wrong response. response can only be n or y", fg='yellow', bold=True)
+
         else:
-          print("Room name can not be empty")
+          click.secho("Room name can not be empty", fg='red', bold=True)
+
 
 
     def do_add_person(self, args):
@@ -97,9 +112,16 @@ class App(cmd.Cmd):
       lastname = parsed_input["<lastname>"]
 
       if firstname and lastname:
-        core.delete_person(firstname,lastname)
+        response = click.secho("Are you sure you want to delete this person [y/n] n ?", fg='yellow', bold=True)
+        if response == "n" or response == "N":
+          click.secho("Delete Cancelled", fg='cyan', bold=True)
+        elif response == "y" or response == "Y": 
+          core.delete_person(firstname,lastname)
+        else:
+          click.secho("Wrong response. response can only be n or y", fg='yellow', bold=True)
       else:
-        print("Provide firstname and lastname of a person you want to delete")
+       click.secho("Provide firstname and lastname of a person you want to delete", fg='red', bold=True)
+
 
     def do_show_people(self,  args):
         """
@@ -214,115 +236,16 @@ class App(cmd.Cmd):
         file_name = parsed_input["<file_name>"]
         core.load_rooms(file_name)
 
+    def do_quit(self, arg):
+            """Quits out of Interactive Mode."""
+
+            click.secho('Thank you for using Amity Allocation App', fg='red', bold=True)
+            exit()
+
 
 if  __name__ ==  "__main__" :
   app = App()
   app.cmdloop()
 
 
-
-
-'''
-
-from core import *
-from docopt import docopt
-
-if __name__ == '__main__':
-    arguments = docopt(__doc__, version='Amity 1.0')
-    #print(arguments)
-
-    if arguments['add_room']:
-      room_type = str(arguments['--room_type'])
-      room_name = arguments['<room_name>']
-      add_room(room_type,room_name)
-
-
-    if arguments['show_rooms']:
-      show_rooms()
-
-    if arguments['delete_room']:
-      index = arguments['<room_identifier>']
-      if not index is None :
-        index = int(index)
-        if isinstance(index, int):
-          index = (arguments['<room_identifier>'] )
-          response = input(" Are you sure you want to delete this room [y/n] n ?")
-          if response == "n" or response == "N":
-            print('Delete Cancelled')
-          elif response == "y" or response == "Y": 
-            delete_room(index)
-          else:
-            print("Wrong response. response can only be n or y")
-        else:
-          print("Room index should be an interger . Use command <show_rooms> to get the right index")
-      else:
-        print("Room index not provided. Use command <show_rooms> to get the right index")
-
-
-    if arguments['add_person']:
-      typ = arguments['--person_type']
-      fname = arguments['<firstname>']
-      lname  = arguments['<lastname>']
-      accomodation = arguments['--wants_accomodation']
-      add_person(typ, fname,lname,accomodation)
-
-
-    if arguments['delete_person']:
-      index = arguments['<person_identifier>']
-      if not index is None :
-        index = int(index)
-        if isinstance(index, int):
-          index = (arguments['<person_identifier>'] )
-          response = input(" Are you sure you want to delete this person [y/n] n ?")
-          if response == "n" or response == "N":
-            print('Delete Cancelled')
-          elif response == "y" or response == "Y": 
-            delete_person(index)
-          else:
-            print("Wrong response. response can only be n or y")
-        else:
-          print("Person index should be an interger . Use command <show_people> to get the right index")
-      else:
-        print("Person index not provided. Use command <show_people> to get the right index")
-
-
-    if arguments['show_people']:
-      show_people()
-
-    if arguments['reallocate']:
-      fname = arguments['<firstname>']
-      lname  = arguments['<lastname>']
-      room_name = arguments['<new_room_name>']
-      reallocate(fname,lname,room_name)
- 
-    if arguments['auto_reallocate']:
-      auto_reallocate()
- 
-    if arguments['print_allocations']:
-      filename =  arguments['-o']
-      print_allocations(filename)
-
-    if arguments['print_unallocated']:
-      filename =  arguments['-o']
-      print_unallocated(filename)
-
-    if arguments['print_room']:
-      room_name = arguments['<print_room_name>']
-      filename =  arguments['-o']
-      if room_name is None:
-        print("You must specify a room")
-      else:
-        print_room(room_name)
-
-    if arguments['load_people']:
-      filename = arguments['<people_file>']
-      if not filename is None:
-        load_people(filename)
-      else:
-        print("Specify the file")
-
-    if arguments['load_rooms']:
-      print("Load command")
-
-'''
